+++ R/data-storage-sqlite.R
@@ -56,6 +56,11 @@ DataStorageSQLite <- R6::R6Class( # nolint object_name.
     connect = function(db_path) {
       # Initialize connection with sqlite database
       private$db_con <- odbc::dbConnect(RSQLite::SQLite(), dbname = db_path)
+    },
+
+    read_data = function(date_from, date_to, bucket) {
+      super$read_data(date_from, date_to, bucket) %>%
+        dplyr::mutate(time = lubridate::as_datetime(time))
     }
   )
 )